requirements.txt
google-generativeai>=0.8.3
openai>=1.0.0
pydantic>=2.0.0
mcstatus
requests
flask
python-dotenv
mcrcon
tenacity
